Updates
onodig codes weg gehaald
verschillend files gemaakt voor de soort screen sizes in saved locaties
neiuw test toegevoegd

reviewd by: Herique
