+++ .idea/deploymentTargetSelector.xml
@@ -19,6 +19,9 @@
       <SelectionState runConfigName="useDatabase()">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
+      <SelectionState runConfigName="deleteLocatie()">
+        <option name="selectionMode" value="DROPDOWN" />
+      </SelectionState>
     </selectionStates>
   </component>
 </project>
