+++ app/src/main/java/com/example/cafefinder/data/database/AppDatabase.kt
@@ -19,7 +19,7 @@ abstract class AppDatabase : RoomDatabase() {
                 val instance = Room.databaseBuilder(
                     context.applicationContext,
                     AppDatabase::class.java,
-                    "app_database2"
+                    "locaties_database"
                 ).build()
                 INSTANCE = instance
                 instance
