+++ app/src/main/java/com/example/cafefinder/data/repository/LocatieDao.kt
@@ -12,6 +12,7 @@ interface LocatieDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(locatie: Locatie)
 
+
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertAll(locaties: List<Locatie>)
 
@@ -24,6 +25,8 @@ interface LocatieDao {
     @Query("DELETE FROM locaties WHERE id = :locatieId")
     suspend fun deleteLocatieById(locatieId: String)
 
+    @Query("SELECT * FROM locaties WHERE id = :id")
+    suspend fun getLocatieById(id: String): Locatie?
 
 
 }
