+++ app/src/main/java/com/example/cafefinder/data/service/SyncService.kt
@@ -1,6 +1,7 @@
 package com.example.cafefinder.data.service
 
 import android.content.Context
+import com.example.cafefinder.data.database.AppDatabase
 import com.example.cafefinder.data.database.LocatieStore
 import com.example.cafefinder.data.model.Locatie
 import kotlinx.coroutines.CoroutineScope
@@ -10,7 +11,7 @@ import kotlinx.coroutines.launch
 class SyncService (private val context: Context ){
 
     private val locatieStore = LocatieStore(context)
-
+    private val locatieDao = AppDatabase.getDatabase(context).locatieDao()
 
     suspend fun syncSaveLocaties(locatie: Locatie) {
         locatieStore.saveLocatie(locatie).collect{ documentId ->
@@ -45,22 +46,15 @@ class SyncService (private val context: Context ){
         CoroutineScope(Dispatchers.IO).launch {
             locatieStore.getAllLocaties().collect { locaties ->
                 locaties.forEach { locatie ->
-                    locatieStore.saveLocatieToRoom(locatie)
+                    val existingLocatie = locatieDao.getLocatieById(locatie.id)
+                    if (existingLocatie == null) {
+                        locatieDao.insert(locatie)
+                    }
                 }
             }
         }
 
     }
 
-
-    suspend fun syncLocatiesFromRoomToFirebase() {
-        CoroutineScope(Dispatchers.IO).launch {
-            val locaties = locatieStore.getAllLocatiesFromRoom()
-            locaties.forEach { locatie ->
-                locatieStore.saveLocatie(locatie)
-            }
-        }
-
-    }
 
 }
